Moved interfaces to the models folder and updated imports across the app
