+++ query-connector/src/app/(pages)/queryBuilding/querySelection/QuerySelection.tsx
@@ -14,7 +14,7 @@ import MyQueriesDisplay from "./QueryLibrary";
 import { SelectedQueryDetails, SelectedQueryState } from "./utils";
 import { BuildStep } from "@/app/shared/constants";
 import { DataContext } from "@/app/shared/DataProvider";
-import { CustomUserQuery } from "@/app/shared/constants";
+import { CustomUserQuery } from "@/app/models/entities/query";
 
 type QuerySelectionProps = {
   selectedQuery: SelectedQueryState;
