+++ query-connector/src/app/models/entities/concepts.ts
@@ -1,0 +1,5 @@
+export interface Concept {
+  code: string;
+  display: string;
+  include: boolean;
+}
