+++ query-connector/src/app/models/entities/query.ts
@@ -1,0 +1,8 @@
+import { DibbsValueSet } from "./valuesets";
+
+export interface CustomUserQuery {
+  query_id: string;
+  query_name: string;
+  conditions_list?: string[];
+  valuesets: DibbsValueSet[];
+}
