+++ query-connector/src/app/shared/constants.ts
@@ -1,4 +1,5 @@
 import { FhirResource } from "fhir/r4";
+import { DibbsConceptType } from "../models/entities/valuesets";
 
 export const USE_CASE_DETAILS = {
   "newborn-screening": {
@@ -146,35 +147,8 @@ export const metadata = {
   description: "Try out TEFCA with queries for public health use cases.",
 };
 
-/*
- * The expected type of a ValueSet concept.
- */
-export interface Concept {
-  code: string;
-  display: string;
-  include: boolean;
-}
-
-/*
- * The expected type of a DIBBS ValueSet.
- */
-export interface DibbsValueSet {
-  valueSetId: string;
-  valueSetVersion: string;
-  valueSetName: string;
-  valueSetExternalId?: string;
-  author: string;
-  system: string;
-  ersdConceptType?: string;
-  dibbsConceptType: DibbsConceptType;
-  includeValueSet: boolean;
-  concepts: Concept[];
-  conditionId?: string;
-}
-
 export const DEFAULT_ERSD_VERSION = "3";
 
-export type DibbsConceptType = "labs" | "conditions" | "medications";
 export type ErsdConceptType =
   | "ostc"
   | "lotc"
@@ -237,11 +211,3 @@ export const HL7_BODY_MISFORMAT =
   "Invalid HL7 request. Please add your HL7 message to the request body in between curly braces like so - { YOUR MESSAGE HERE } ";
 export const MISSING_PATIENT_IDENTIFIERS =
   "No patient identifiers to parse from requestBody.";
-
-// Type definition for CustomUserQueries
-export interface CustomUserQuery {
-  query_id: string;
-  query_name: string;
-  conditions_list?: string[];
-  valuesets: DibbsValueSet[];
-}
